fix: swap countly events with sa
