+++ packages/website/components/tokens/tokenCreator/tokenCreator.js
@@ -3,7 +3,7 @@ import clsx from 'clsx';
 import { useRouter } from 'next/router';
 import { useCallback, useEffect, useLayoutEffect, useRef, useState } from 'react';
 
-import countly from 'lib/countly';
+import analytics from 'lib/analytics';
 import Button, { ButtonVariant } from 'components/button/button';
 import { useTokens } from 'components/contexts/tokensContext';
 import { useUser } from 'hooks/use-user';
@@ -30,15 +30,16 @@ const TokenCreator = ({ content }) => {
   const onTokenCreate = useCallback(
     async e => {
       // Tracking
-      countly.trackEvent(
-        countly.events.TOKEN_CREATE,
+      // @ts-ignore
+      window?.sa_event(
+        analytics.events.TOKEN_CREATE,
         !tokens.length
           ? {
-              ui: countly.ui.TOKENS_EMPTY,
+              ui: analytics.ui.TOKENS_EMPTY,
               action: 'New API Token',
             }
           : {
-              ui: countly.ui.NEW_TOKEN,
+              ui: analytics.ui.NEW_TOKEN,
               action: 'Create new token',
             }
       );
