+++ packages/website/pages/404.js
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 
-import countly from '../lib/countly';
+import analytics from '../lib/analytics';
 import GeneralPageData from '../content/pages/general.json';
 import BlockBuilder from '../components/blockbuilder/blockbuilder.js';
 
@@ -8,7 +8,8 @@ export default function Home() {
   const blocks = GeneralPageData.error_404.sections;
 
   useEffect(() => {
-    countly.trackEvent(countly.events.NOT_FOUND, {
+    // @ts-ignore
+    window?.sa_event(analytics.events.NOT_FOUND, {
       path: window.location?.pathname ?? 'unknown',
       referrer: typeof window !== 'undefined' ? document.referrer : null,
     });
