+++ packages/website/pages/_app.js
@@ -28,8 +28,12 @@ const App = ({ Component, pageProps }) => {
     <AppProviders authorizationProps={{ ...pageProps }}>
       <Metadata {...pageProps} />
       <RestrictedRoute {...pageProps}>
-        <Script src="https://scripts.simpleanalyticscdn.com/latest.dev.js" data-hostname="web3.storage" />
-        <Script src="https://scripts.simpleanalyticscdn.com/auto-events.js" />
+        <Script
+          src="https://track.web3.storage/latest.dev.js"
+          data-hostname="web3.storage"
+          data-ignore-pages="/callback"
+        />
+        <Script src="https://track.web3.storage/auto-events.js" />
         <noscript>
           {/* eslint-disable @next/next/no-img-element */}
           <img
